reader/requirements.txt
requests
RPi.GPIO
spi

